style: refactors the NotAvailableMessage component
Now you can use it from all the front
Fixs the writing of the word available
