+++ src/pages/Booth/Panel/CastVoteView.jsx
@@ -4,7 +4,7 @@ import { Button } from "react-bulma-components";
 import { useParams, useSearchParams } from "react-router-dom";
 import { getVotesInfo } from "../../../services/info";
 import UrnaTable from "./components/UrnaTable";
-import NotAvalaibleMessage from "../components/NotAvalaibleMessage";
+import NotAvalaibleMessage from "../../../component/Messages/NotAvailableMessage";
 
 function CastVoteView({ election }) {
   const [electionData, setElectionData] = useState({
